added contests page
